+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Possibility to set the attribute that the user has to change his password at the next login (#51)
 
 ## [v1.5.0-3] - 2022-04-26
 ### Removed